ref(endpoint): Separation of responsibility for modules
- The endpoint `users/create/` is now moved to `app/api/users.py`.
- The create method (C.R.U.D.) for User model is now moved to `app/api/crud`.
